PHPStan big cleaning

+++ themes/Rozier/Forms/UserDetailsType.php
@@ -107,7 +107,7 @@ class UserDetailsType extends AbstractType
                     'day' => 'day'
                 ],
                 'required' => false,
-                'years' => range(1920, date('Y') - 6),
+                'years' => range(1920, ((int) date('Y')) - 6),
                 'widget' => 'single_text',
                 'format' => 'yyyy-MM-dd',
                 'attr' => [
